test(py): simplified test ros_brdige
