+++ mosaico-sdk-py/src/testing/unit/ros_bridge/adapters/test_sensor_msgs.py
@@ -71,14 +71,6 @@ ROS_TYPESTORE_TO_TEST = [
 ]
 
 
-@pytest.fixture
-def invalid_ms_msg() -> Message:
-    return Message(
-        data=Serializable(),
-        timestamp_ns=0,
-    )
-
-
 ###############################################################################
 ########################### TestCameraInfoAdapter #############################
 ###############################################################################
